api throttle updated

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
@@ -13,9 +14,12 @@ class PostController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return PostResource::collection(Post::with('author')->paginate(2));
+
+        $user = request()->user();
+        $posts = $user->posts()->paginate(2);
+        return PostResource::collection($posts);
     }
 
     /**
@@ -24,7 +28,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id;
         $post = Post::create($data);
         return new PostResource($post);
     }
@@ -34,6 +38,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
         $post->load('author');
         return new PostResource($post);
     }
@@ -43,6 +48,7 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
         $data = $request->validated();
 
         $post->update($data);
@@ -56,6 +62,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden');
         $post->delete();
 
         return response()->noContent();
